requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dataclasses-json==0.6.1
decorator==4.4.2
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.103.2
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==0.18.3
greenlet==3.0.1
h11==0.14.0
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
jwt==1.3.1
langchain==0.0.332
langsmith==0.0.62
marshmallow==3.20.1
motor==3.3.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
pluggy==1.3.0
proglog==0.1.10
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.6.0
pytest==7.4.3
pytest-subtests==0.11.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2
